refactor: switch config to direct self-reexport imports
