+++ packages/opencode/src/cli/cmd/providers.ts
@@ -7,7 +7,7 @@ import { ModelsDev } from "../../provider"
 import { map, pipe, sortBy, values } from "remeda"
 import path from "path"
 import os from "os"
-import { Config } from "../../config"
+import { Config } from "@/config/config"
 import { Global } from "../../global"
 import { Plugin } from "../../plugin"
 import { Instance } from "../../project/instance"
