+++ packages/opencode/src/config/index.ts
@@ -1,3 +1,2 @@
-export * as Config from "./config"
 export * as ConfigMarkdown from "./markdown"
 export * as ConfigPaths from "./paths"
