+++ packages/opencode/src/server/instance/experimental.ts
@@ -8,7 +8,7 @@ import { Instance } from "../../project/instance"
 import { Project } from "../../project"
 import { MCP } from "../../mcp"
 import { Session } from "../../session"
-import { Config } from "../../config"
+import { Config } from "@/config/config"
 import { ConsoleState } from "../../config/console-state"
 import { Account, AccountID, OrgID } from "../../account"
 import { AppRuntime } from "../../effect/app-runtime"
