+++ packages/opencode/src/share/share-next.ts
@@ -10,7 +10,7 @@ import { Session } from "@/session"
 import { MessageV2 } from "@/session/message-v2"
 import type { SessionID } from "@/session/schema"
 import { Database, eq } from "@/storage"
-import { Config } from "@/config"
+import { Config } from "@/config/config"
 import { Log } from "@/util"
 import { SessionShareTable } from "./share.sql"
