+++ packages/opencode/src/skill/skill.ts
@@ -11,7 +11,7 @@ import { Flag } from "@/flag/flag"
 import { Global } from "@/global"
 import { Permission } from "@/permission"
 import { AppFileSystem } from "@opencode-ai/shared/filesystem"
-import { Config } from "../config"
+import { Config } from "@/config/config"
 import { ConfigMarkdown } from "../config"
 import { Glob } from "@opencode-ai/shared/util/glob"
 import { Log } from "../util"
