+++ packages/opencode/src/snapshot/snapshot.ts
@@ -7,7 +7,7 @@ import * as CrossSpawnSpawner from "@/effect/cross-spawn-spawner"
 import { InstanceState } from "@/effect"
 import { AppFileSystem } from "@opencode-ai/shared/filesystem"
 import { Hash } from "@opencode-ai/shared/util/hash"
-import { Config } from "../config"
+import { Config } from "@/config/config"
 import { Global } from "../global"
 import { Log } from "../util"
