+++ packages/opencode/test/share/share-next.test.ts
@@ -8,7 +8,7 @@ import { Account } from "../../src/account"
 import { AccountRepo } from "../../src/account/repo"
 import * as CrossSpawnSpawner from "../../src/effect/cross-spawn-spawner"
 import { Bus } from "../../src/bus"
-import { Config } from "../../src/config"
+import { Config } from "../../src/config/config"
 import { Provider } from "../../src/provider"
 import { Session } from "../../src/session"
 import type { SessionID } from "../../src/session/schema"
